Extract `RegexNFAState` and tagged transition classes into their own files. (#47)
