+++ CMakeLists.txt
@@ -97,6 +97,9 @@ set(SOURCE_FILES
     src/log_surgeon/finite_automata/RegexDFA.hpp
     src/log_surgeon/finite_automata/RegexDFA.tpp
     src/log_surgeon/finite_automata/RegexNFA.hpp
+    src/log_surgeon/finite_automata/RegexNFAState.hpp
+    src/log_surgeon/finite_automata/RegexNFAStateType.hpp
+    src/log_surgeon/finite_automata/TaggedTransition.hpp
     src/log_surgeon/finite_automata/UnicodeIntervalTree.hpp
     src/log_surgeon/finite_automata/UnicodeIntervalTree.tpp
     )
